reconstruction/PBIDR/requirements.txt
aadict==0.2.3
absl-py==0.12.0
aiohttp==3.7.4.post0
ansicolors==1.1.8
ansiwrap==0.8.4
apptools==5.1.0
argon2-cffi==20.1.0
asset==0.6.13
astor==0.8.1
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
autoprop==3.0.0
backcall==0.2.0
bleach==3.3.0
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.5
chainmap==1.0.3
chardet==4.0.0
charset-normalizer==2.0.4
chumpy==0.70
cloudpickle==1.6.0
colorlog==5.0.1
combomethod==1.0.12
ConfigArgParse==1.4.1
configobj==5.0.6
constantly==15.1.0
cryptography==35.0.0
cvxpy==1.1.13
cycler==0.10.0
Cython==0.29.24
dask==2021.7.0
decorator==4.4.2
defusedxml==0.7.1
ecos==2.0.7.post1
entrypoints==0.3
envisage==6.0.1
face-alignment==1.3.3
freetype-py==2.2.0
fsspec==2021.7.0
future==0.18.2
fvcore==0.1.5.post20210423
gast==0.4.0
glcontext==2.3.4
globre==0.1.5
glooey==0.3.5
google-pasta==0.2.0
GPUtil==1.4.0
grpcio==1.37.1
h5py==3.2.1
hyperlink==21.0.0
idna==3.2
imageio==2.9.0
imageio-ffmpeg==0.4.3
importlib-metadata==4.0.1
importlib-resources==5.3.0
incremental==21.3.0
intspan==1.6.1
iopath==0.1.8
ipykernel==5.5.3
ipython==7.23.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter-client==6.2.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
llvmlite==0.36.0
locket==0.2.1
lxml==4.6.3
mapbox-earcut==0.12.10
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.1
matplotlib-inline==0.1.2
mayavi==4.7.3
mementos==1.3.1
menpo==0.11.0
menpo3d==0.8.3
meshio==4.4.6
mistune==0.8.4
moderngl==5.6.4
more-itertools==8.8.0
mpmath==1.2.1
msgpack==1.0.2
multidict==5.2.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
notebook==6.3.0
nulltype==2.3.1
numba==0.53.1
numpy==1.20.2
opencv-python==4.5.1.48
options==1.4.10
osqp==0.6.2.post0
packaging==20.9
pandas==1.3.5
pandocfilters==1.4.3
parso==0.8.2
partd==1.2.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
plotly==5.1.0
portalocker==2.3.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.17.0
psutil==5.8.0
ptyprocess==0.7.0
pycollada==0.7.1
pycparser==2.20
pyembree==0.1.6
pyface==7.3.0
pyglet==1.5.18
Pygments==2.9.0
pyhocon==0.3.57
PyOpenGL==3.1.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyrender==0.1.45
pyrsistent==0.17.3
python-dateutil==2.8.1
python-fcl==0.0.12
pytorch3d==0.4.0
pytz==2021.3
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.0.3
qdldl==0.1.5.post0
readline==6.2.4.1
requests==2.26.0
retrying==1.3.3
Rtree==0.9.7
say==1.6.6
scikit-image==0.18.1
scipy==1.6.2
scs==2.1.3
Send2Trash==1.5.0
Shapely==1.7.1
show==1.6.0
signature-dispatch==0.1.0
simplere==1.2.13
six==1.12.0
svg.path==4.1
sympy==1.8
tabulate==0.8.9
tenacity==7.0.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
textdata==2.4.1
textwrap3==0.9.2
tifffile==2021.4.8
toolz==0.11.1
torch==1.7.1
torchvision==0.7.0
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
traits==6.3.1
traitsui==7.2.1
triangle==20200424
trimesh==3.9.25
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.7.4.3
urllib3==1.26.6
vecrec==0.3.0
vtk==9.0.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
wrapt==1.12.1
wslink==1.1.0
xatlas==0.0.5
xxhash==2.0.2
yacs==0.1.8
yarl==1.7.0
zipp==3.4.1
zope.interface==5.4.0
